chore: refactor domain types
