+++ src/Domain/Decision.ts
@@ -1,0 +1,6 @@
+export type Decision<State> = {
+  allowed: boolean;
+  retryAfter: number;
+  remaining: number;
+  nextState: State;
+};
