+++ src/Domain/Policies.ts
@@ -1,0 +1,5 @@
+export type Policies<Policy> = {
+  apiKey: Policy;
+  ip?: Policy;
+  tenant?: Policy;
+};
