+++ src/Domain/Service/RateLimiterService.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, test, vi } from "vitest";
 import { StateRepositoryInterface } from "../Repository/StateRepositoryInterface.ts";
-import { Decision, Policies, StateIdentifiers } from "../types.ts";
 import { RateLimiterService } from "./RateLimiterService.ts";
 import { RateLimiterInterface } from "../Algorithm/RateLimiterInterface.ts";
 
@@ -21,7 +20,7 @@ describe("limit service", () => {
     };
     const expectedDecision = {
       nextState: { key: "new-state-key" },
-    } as Decision<FakeState>;
+    };
     const MockedAlgorithm: RateLimiterInterface<FakeState, FakeConfig> = {
       attempt: vi.fn().mockReturnValue(expectedDecision),
     };
@@ -29,8 +28,8 @@ describe("limit service", () => {
       mockedRepository,
       MockedAlgorithm,
     );
-    const stateIdentifiers: StateIdentifiers = { apiKey: "apikey-identifier" };
-    const algorithmConfig: Policies<FakeConfig> = {
+    const stateIdentifiers = { apiKey: "apikey-identifier" };
+    const algorithmConfig = {
       apiKey: { key: "example-config-key" },
     };
 
@@ -64,11 +63,11 @@ describe("limit service", () => {
     const expectedDecisions = {
       apiKey: {
         nextState: { key: "new-apiKey-state-key" },
-      } as Decision<FakeState>,
-      ip: { nextState: { key: "new-ip-state-key" } } as Decision<FakeState>,
+      },
+      ip: { nextState: { key: "new-ip-state-key" } },
       tenant: {
         nextState: { key: "new-tenant-state-key" },
-      } as Decision<FakeState>,
+      },
     };
     const MockedAlgorithm: RateLimiterInterface<FakeState, FakeConfig> = {
       attempt: vi
@@ -81,12 +80,12 @@ describe("limit service", () => {
       mockedRepository,
       MockedAlgorithm,
     );
-    const stateIdentifiers: StateIdentifiers = {
+    const stateIdentifiers = {
       apiKey: "apikey-identifier",
       ip: "ip-identifier",
       tenant: "tenant-identifier",
     };
-    const algorithmConfig: Policies<FakeConfig> = {
+    const algorithmConfig = {
       apiKey: { key: "api-config" },
       ip: { key: "ip-config" },
       tenant: { key: "tenant-config" },
