+++ src/Domain/StateIdentifiers.ts
@@ -1,0 +1,5 @@
+export type StateIdentifiers = {
+  apiKey: string;
+  ip?: string;
+  tenant?: string;
+};
